+++ src/mock.rs
@@ -2,9 +2,10 @@ use std::collections::HashMap;
 
 use snafu::ResultExt;
 
+use crate::encoding::Binary;
 use crate::errors::{ContractErr, Result, Utf8StringErr};
 use crate::traits::{Api, Extern, ReadonlyStorage, Storage};
-use crate::types::{BlockInfo, CanonicalAddr, Coin, ContractInfo, HumanAddr, MessageInfo, Params};
+use crate::types::{BlockInfo, CanonicalAddr, Coin, ContractInfo, Env, HumanAddr, MessageInfo};
 
 // dependencies are all external requirements that can be injected for unit tests
 pub fn dependencies(canonical_length: usize) -> Extern<MockStorage, MockApi> {
@@ -78,13 +79,13 @@ impl Api for MockApi {
         if append > 0 {
             out.extend(vec![0u8; append]);
         }
-        Ok(CanonicalAddr(out))
+        Ok(CanonicalAddr(Binary(out)))
     }
 
     fn human_address(&self, canonical: &CanonicalAddr) -> Result<HumanAddr> {
         // remove trailing 0's (TODO: fix this - but fine for first tests)
         let trimmed: Vec<u8> = canonical
-            .as_bytes()
+            .as_slice()
             .iter()
             .cloned()
             .filter(|&x| x != 0)
@@ -97,14 +98,14 @@ impl Api for MockApi {
 
 // just set signer, sent funds, and balance - rest given defaults
 // this is intended for use in testcode only
-pub fn mock_params<T: Api, U: Into<HumanAddr>>(
+pub fn mock_env<T: Api, U: Into<HumanAddr>>(
     api: &T,
     signer: U,
     sent: &[Coin],
     balance: &[Coin],
-) -> Params {
+) -> Env {
     let signer = signer.into();
-    Params {
+    Env {
         block: BlockInfo {
             height: 12_345,
             time: 1_571_797_419,
@@ -138,14 +139,14 @@ mod test {
     use crate::types::coin;
 
     #[test]
-    fn mock_params_arguments() {
+    fn mock_env_arguments() {
         let name = HumanAddr("my name".to_string());
         let api = MockApi::new(20);
 
         // make sure we can generate with &str, &HumanAddr, and HumanAddr
-        let a = mock_params(&api, "my name", &[], &coin("100", "atom"));
-        let b = mock_params(&api, &name, &[], &coin("100", "atom"));
-        let c = mock_params(&api, name, &[], &coin("100", "atom"));
+        let a = mock_env(&api, "my name", &[], &coin("100", "atom"));
+        let b = mock_env(&api, &name, &[], &coin("100", "atom"));
+        let c = mock_env(&api, name, &[], &coin("100", "atom"));
 
         // and the results are the same
         assert_eq!(a, b);
@@ -167,8 +168,8 @@ mod test {
         let human = HumanAddr("shorty".to_string());
         let canon = api.canonical_address(&human).unwrap();
         assert_eq!(canon.len(), 20);
-        assert_eq!(&canon.as_bytes()[0..6], human.as_str().as_bytes());
-        assert_eq!(&canon.as_bytes()[6..], &[0u8; 14]);
+        assert_eq!(&canon.as_slice()[0..6], human.as_str().as_bytes());
+        assert_eq!(&canon.as_slice()[6..], &[0u8; 14]);
 
         let recovered = api.human_address(&canon).unwrap();
         assert_eq!(human, recovered);